# faster Array.averageBy, and some cleanup