This re-implements the improved averageBy from #886 which appears to have gotten stalled.
I also did some extraneous trailing semi colon and paren removal

If this gets merged you can also close out #886 

Performance comparison with (fun x -> x*x) 


```ini

Host Process Environment Information:
BenchmarkDotNet=v0.9.8.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-4712HQ CPU 2.30GHz, ProcessorCount=8
Frequency=2240908 ticks, Resolution=446.2477 ns, Timer=TSC
CLR=MS.NET 4.0.30319.42000, Arch=64-bit RELEASE [RyuJIT]
GC=Concurrent Workstation
JitModules=clrjit-v4.6.1590.0

Type=SIMDBenchmark  Mode=Throughput  Platform=X64  
Jit=RyuJit  GarbageCollection=Concurrent Workstation  

```
 Method |  Length |            Median |          StdDev | Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
------- |-------- |------------------ |---------------- |------ |------ |------ |------------------- |
    **Old** |      **10** |       **116.1816 ns** |       **1.8323 ns** |  **0.01** |     **-** |     **-** |              **13.95** |
    New |      10 |         6.7990 ns |       0.4203 ns |     - |     - |     - |               0.00 |
    **Old** |   **10000** |    **64,502.2371 ns** |     **859.8698 ns** |     **-** |     **-** |     **-** |              **39.07** |
    New |   10000 |     9,675.1776 ns |     122.1936 ns |     - |     - |     - |               3.41 |
    **Old** | **1000000** | **6,533,337.9930 ns** | **228,221.1828 ns** |     **-** |     **-** |  **1.00** |         **228,818.56** |
    New | 1000000 | 1,076,279.6822 ns |  12,940.7151 ns |     - |     - |  0.12 |          29,960.47 |