@forki 2016-08-15
I think it's fine for now, but please send the comma thingy as a separate PR (next time)